+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # asma-core-ui
 
+## 1.6.0
+
+### Minor Changes
+
+-   37e14dc: StyledDataGrid
+
+### Patch Changes
+
+-   37e14dc: DataGrid: Row background, custom checkbox classes
+
+## 1.5.9
+
+### Patch Changes
+
+-   58f8d13: fixed height of styledDialog closeIcon
+
+## 1.5.7
+
+### Patch Changes
+
+-   601522c: added showCloseIcon props to StyledDialog
+
 ## 1.5.3
 
 ### Patch Changes